Create dataset search page and move config provider out of entity search
