+++ context/app/static/js/components/entitySearch/EntitySearch/style.js
