+++ context/app/static/js/pages/entitySearch/DatasetSearch/index.js
@@ -1,0 +1,3 @@
+import DatasetSearch from './DatasetSearch';
+
+export default DatasetSearch;
